message_view: Allow quoting selected message text, via popover option.
Whenever the user selects text within a single message, a tooltip, with
the option to quote and reply, will appear next to the selection.
Clicking it will quote the markdown version of the selection, hence
preserving any formatting, in the compose box, opening it if necessary.

Any other elements of the message, outside of the content, selected
presumably by accident (like the timestamp or sender name), will be
ignored.

The 2 standard ways to quote and reply, via the message actions menu,
and the `>` keyboard shortcut, still behave as before, quoting the whole
message.

Fixes: #19712.